refactor(components.converter): move to webUniverseComponents
