+++ client/app/webUniverseComponents/index.js
@@ -1,8 +1,12 @@
 import angular from 'angular';
 
+import wucConverter from './converter';
+
 const moduleName = 'webUniverseComponents';
 
 angular
-  .module(moduleName, []);
+  .module(moduleName, [
+    wucConverter,
+  ]);
 
 export default moduleName;
